moved Java program into Java specific folder. Same for Python
