--- CS147DV/src/parse/ParseTest.java
@@ -1,18 +1,0 @@
-package parse;
-
-/**
- * 
- * @author Jordan
- *
- *	A class that tests the console capabilities of AssemblyParser
- */
-public class ParseTest
-{
-
-	public static void main(String[] args)
-	{
-		AssemblyParser.startConsoleParser();
-
-	}
-
-}
